[backend]

+++ backend/chat/controller.js
@@ -1,19 +1,14 @@
 //console.log("controller");
 
 
-
-
 import { User } from "../auth/model.js";
-import { Message, Group, GroupMessage, GroupMembers, Story, Friendship } from "./model.js";
+import { Message, Group, GroupMessage, Story, Friendship } from "./model.js";
 import { io, onlineUserMap } from '../utils/socket.js';
 import { ObjectId } from "mongodb";
-
-
+import mongoose from "mongoose";
 
 
 //console.log( "controller", onlineUserMap);
-
-
 
 
 export const fetchMessage = async (req, res, next) => {
@@ -23,7 +18,7 @@ export const fetchMessage = async (req, res, next) => {
 		const { id } = req.body;
 
 		const partner = await User.findOne({
-			_id: new ObjectId( id )
+			_id: new ObjectId(id)
 		})
 
 
@@ -89,6 +84,61 @@ export const sendMessage = async (req, res) => {
 	}
 };
 
+export const sendGroupMessage = async (req, res) => {
+	try {
+		console.dir("send group message #########################################");
+		//console.dir(req.body)
+
+
+		const { receiver, messages } = req.body;
+
+		//console.log( receiver, messages )
+
+		const saved_messages = []
+
+		console.log("here")
+
+		for (const elem of messages) {
+			const newMessage = await GroupMessage.create({
+				sender_id: req.user_id,
+				group_id: receiver._id,
+				type: elem.type,
+				content: elem.content
+			})
+
+
+			const newMessage2 = await GroupMessage.findOne({ _id: newMessage._id }).populate("sender_id", "username _id photo");
+
+			//console.log(newMessage)
+
+			saved_messages.push(newMessage2)
+		}
+
+		console.log("here2")
+
+		console.log("saved_messages", saved_messages);
+
+		for (let member of receiver.members) {
+			console.log(member)
+			if (member.username === req.username) continue;
+			let receiver_socket_id = onlineUserMap[member.username];
+			if (receiver_socket_id) {
+				console.log("emitting message to socket:", receiver_socket_id, receiver);
+				io.to(receiver_socket_id).emit("receive_group_message", { messages: saved_messages });
+			}
+		}
+
+
+
+		return res.status(200).json({ messages: saved_messages });
+
+	} catch (err) {
+		console.error(err);
+		return res.status(400).json({ error: err.message });
+	}
+};
+
+
 
 
 export const FetchUsers = async (req, res, next) => {
@@ -111,38 +161,61 @@ export const CreateGroup = async (req, res, next) => {
 		const { newGroup } = req.body
 		console.log(newGroup)
 
-		const new_group = new Group({
-			admin: req.username,
+		const new_group = await Group.create({
+			admin: req.user_id,
 			name: newGroup,
+			members: [req.user_id]
 		})
 
-		const saved_group = await new_group.save();
 
-		const me = await User.findOne({ username: req.username })
-
-		const new_member = new GroupMembers({
-			group_id: saved_group._id,
-			group_name: saved_group.name,
-			member: req.username,
-			photo: me.photo,
-			admin: req.username
-		})
-
-		const saved_member = await new_member.save();
-
-		res.status(200).json(saved_member)
+		res.status(200).json({ new_group })
 	} catch (err) {
+
+		if (err.code === 11000) {
+			const field = Object.keys(err.keyValue)[0];
+			return res.status(400).json({ error: "You already have a group of this name" })
+		}
 		res.status(400).json({ error: err.message })
 	}
 }
 
 
 export const FetchGroups = async (req, res, next) => {
-	console.log("fetch groups")
+	console.log("fetch groups #############################################");
+	console.log(req.user_id);
 	try {
-		const groups = await GroupMembers.find({ member: req.username })
-		//console.log(groups)
-		res.status(200).json(groups)
+		let { page, limit, searchFor } = req.query;
+		page = parseInt(page);
+		limit = parseInt(limit);
+
+		const userId = req.user_id;
+
+		const groups = await Group.aggregate([
+			{
+				$match: {
+					members: req.user_id,
+					name: { $regex: searchFor, $options: "i" }
+				}
+			},
+			{
+				$facet: {
+					data: [
+						{ $skip: (page - 1) * limit },
+						{ $limit: limit }
+					],
+					pages: [
+						{ $count: "count" }
+					]
+				}
+			}
+		]);
+
+		const data = groups[0]?.data || [];
+		const total = groups[0]?.pages[0]?.count || 0;
+		const pages = Math.ceil(total / limit);
+
+
+		res.status(200).json({ data, pages });
 	} catch (err) {
 		res.status(400).json({ error: err.message })
 	}
@@ -152,22 +225,19 @@ export const FetchGroups = async (req, res, next) => {
 export const AddToGroup = async (req, res, next) => {
 	console.log("add to group")
 	try {
-		const { new_member, group_name } = req.body
-		console.log(new_member, group_name)
+		let { new_member, group } = req.body
+		console.log(new_member, group)
 
-		const group = await Group.findOne({ name: group_name, admin: req.username })
-		const nmember = await User.findOne({ username: new_member })
-		const new_data = new GroupMembers({
-			group_id: group._id,
-			group_name: group.name,
-			member: new_member,
-			photo: nmember.photo,
-			admin: req.username
-		})
+		await Group.updateOne(
+			{ _id: group._id },
+			{
+				$addToSet: { members: new_member._id }
+			}
+		)
 
-		const saved_data = await new_data.save()
 
-		res.status(200).json(saved_data)
+
+		res.status(200).json({})
 	} catch (err) {
 		res.status(400).json({ error: err.message })
 		console.log(err)
@@ -176,16 +246,22 @@ export const AddToGroup = async (req, res, next) => {
 
 
 export const DeleteFromGroup = async (req, res, next) => {
-	console.log("delete from group")
+	console.log("remove from group")
 	try {
-		const { group_id, member } = req.body
-		await GroupMembers.deleteOne({ group_id, member })
-		res.status(200).json("ok")
+		let { member, group } = req.body
+		console.log(member, group)
+
+		await Group.updateOne(
+			{ _id: group._id },
+			{
+				$pull: { members: member._id }
+			}
+		);
+
+		res.status(200).json({})
 	} catch (err) {
 		res.status(400).json({ error: err.message })
-	}
-	finally {
-		next()
+		console.log(err)
 	}
 }
 
@@ -201,41 +277,77 @@ export const FetchGroupMembers = async (req, res, next) => {
 	}
 }
 
-export const DeleteGroup = async (req, res, next) => {
-	console.log("delete group")
-	try {
-		const { group_id } = req.body
-		await Group.deleteOne({ _id: group_id })
-		await GroupMembers.deleteMany({ group_id })
-		await GroupMessage.deleteMany({ group_id })
-		res.status(200).json("ok")
-	} catch (err) {
-		res.status(400).json({ error: err.message })
-	}
-}
 
 
 export const FetchGroupMessage = async (req, res, next) => {
 	console.log("fetch group message");
+
 	try {
-		const { group_id } = req.body
-		const messages = await GroupMessage.find({ group_id })
+		const { id: group_id } = req.params;
+		const group = await Group.findOne({ _id: group_id }).populate("members", "name _id username photo").populate("admin", "name _id username photo");
+		const messages = await GroupMessage.find({ group_id }).populate("sender_id", "username _id photo");
 		//console.log(messages)
-		res.status(200).json(messages)
+		res.status(200).json({ messages, partner: group });
 	} catch (err) {
 		res.status(400).json({ error: err.message })
 	}
 }
 
+export const DeleteGroup = async (req, res, next) => {
+	console.log("delete group")
+	try {
+		const { group_id } = req.body;
+
+		if (!mongoose.Types.ObjectId.isValid(group_id)) {
+			return res.status(400).json({ error: "Invalid group id" });
+		}
+
+		const group = await Group.findById(group_id);
+
+		if (!group) {
+			return res.status(404).json({ error: "Group not found" });
+		}
+
+		await Group.deleteOne({ _id: group_id })
+		await GroupMessage.deleteMany({ group_id })
+		res.status(200).json({ message: "Success" })
+	} catch (err) {
+		console.log(err.message)
+		res.status(400).json({ error: err.message })
+	}
+}
 
 
 export const LeaveGroup = async (req, res, next) => {
 	console.log("leave group")
 	try {
 		const { group_id } = req.body
-		await GroupMembers.deleteOne({ member: req.username });
-		res.status(200).json("Left From The Group")
+
+		if (!mongoose.Types.ObjectId.isValid(group_id)) {
+			return res.status(400).json({ error: "Invalid group id" });
+		}
+
+		const group = await Group.findById(group_id);
+
+		if (!group) {
+			return res.status(404).json({ error: "Group not found" });
+		}
+
+		if (group.admin.equals(req.user_id)) {
+			return res.status(400).json({
+				error: "Admin cannot leave. Transfer ownership or delete group."
+			});
+		}
+
+		const result = await Group.updateOne(
+			{ _id: group_id },
+			{ $pull: { members: req.user_id } }
+		);
+
+		
+		res.status(200).json({ message: "Success" })
 	} catch (err) {
+		console.log(err.message);
 		res.status(400).json({ error: err.message })
 	}
 }
